Audit-driven doc + code fixes (32-agent review)
Docs (EN+JA):
- Fix IRC step-size: "mass-weighted" → "unweighted Bohr"
- Fix DFT basis: def2-TZVPD → def2-tzvpd (match code)
- Fix IRC: energy-based predictor (not steepest-descent)
- Fix --include-H2O/--add-linkH → lowercase
- Fix "for the ML region" → removed (pure MLIP, no ONIOM)
- Fix DFT//UMA → DFT//MLIP in JA
- Fix return_partial_hessian: false → true in JA yaml_reference
- Add print_timing to uma_pysis config reference
- Fix non-breaking hyphens in getting_started.md
- JA: 一気通貫 (end-to-end), 虚振動数 consistency

Code:
- Unify --preopt default to False (all, path-search, scan*)
- Fix path_search.py RMSD unit "Å" → "bohr"
- Fix scan.py docstring: remove "using UMA calculator"
- Remove "for the ML region" from multiplicity help (9 files)
- Fix xtb_alpb_correction.py normalize_solvent_name .lower()
- Fix orb.py charge/spin float→int
- Fix aimnet2.py remove "gradient" from forces fallback
- Fix uma.py reshape before _zero_frozen_forces_ev
- Fix extract.py exclude_backbone default mismatch
- Fix EulerPC.py hessian_recalc: slice to active DOFs

+++ pdb2reaction/backends/orb.py
@@ -183,8 +183,8 @@ class OrbCalculator(MLIPCalculator):
         from ase import Atoms
 
         atoms = Atoms(symbols=list(elem), positions=np.asarray(coord_ang, dtype=np.float64))
-        atoms.info["charge"] = float(self.charge)
-        atoms.info["spin"] = float(self.mult)
+        atoms.info["charge"] = int(self.charge)
+        atoms.info["spin"] = int(self.mult)
         atoms.calc = self._ase_calc
 
         energy = float(atoms.get_potential_energy())
